+++ NEWS.md
@@ -32,8 +32,6 @@
 
 * Added/updated functions `add_n()` and `add_nevent()` to work with `tbl_regression` and `tbl_uvregression` objects. Each function now has an argument to place Ns on the label or level rows. (#744)
 
-* Added `"jama_psychiatry"` theme to `theme_gtsummary_journal()` (#789)
-
 * Added _The Quarterly Journal of Economics_ to `theme_gtsummary_journal()`. This journal theme will be updated again after the gt package updates `cols_merge()` with a rows argument and allows for line breaks within cell.
 
 * Added default support for `brmsfit` model in `tbl_regression()` with new method function. (#751)
@@ -68,6 +66,8 @@
 
 * Allowing for tidyverse shortcut notation in `tbl_survfit(estimate_fun=)` specification, e.g. `tbl_survfit(estimate_fun= ~style_sigfig(.x * 100))` (#761)
 
+* The JAMA journal theme has been updated to merge the coefficient and confidence interval columns.
+
 ### Bug Fixes
 
 * The `label=` argument for unstratified models was being ignored in `tbl_survfit()` (#842)
@@ -83,6 +83,10 @@
 ### Breaking Changes
 
 * `tbl_survival()` has moved from questioning to deprecated. This function maintains the old `x$table_header` format (instead of the more flexible `x$table_styling`). The `"level_label"` column was renamed to `"groupname_col"` and the `x$table_body` is no longer grouped with `group_by(level_label)` 
+
+* The back-end implementation of `add_stat_label(location = "row")` has been updated. The function now merges columns `"label"` and `"stat_label"` instead of modifying the `"label"` column directly. This _could_ be a breaking change if users had manually updated results, for example, from a `tbl_regression()` table to merge with `tbl_summary()` using `tbl_merge()` 
+
+* The function `add_stat_label()` no longer auto-switches `location = "column"` requests to `"row"` in the presence of `"continuous2"` variables.
 
 # gtsummary 1.3.7